Refactor: rename scripts to src and create a module
